Update 03 - Refactoring of all files and autoload
