+++ src/App/Repositories/Article.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Repository;
+namespace App\Repositories;
 
 class Article extends Model
 {
