+++ src/App/Repositories/Comment.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Repository;
+namespace App\Repositories;
 
 class Comment extends Model
 {
